add customOption helper
