+++ core/modules/catalog/components/ProductCustomOptions.ts
@@ -1,49 +1,9 @@
+import { customOptionFieldName, selectedCustomOptionValue, defaultCustomOptionValue } from '@vue-storefront/core/modules/catalog/helpers/customOption';
 import { mapMutations } from 'vuex'
 import * as types from '../store/product/mutation-types'
 import rootStore from '@vue-storefront/core/store'
 import i18n from '@vue-storefront/i18n'
 import { Logger } from '@vue-storefront/core/lib/logger'
-
-function _defaultOptionValue (co) {
-  switch (co.type) {
-    case 'radio': {
-      return co.values && co.values.length ? co.values[0].option_type_id : 0
-    }
-    case 'checkbox': {
-      return []
-    }
-    default: {
-      return ''
-    }
-  }
-}
-
-function _fieldName (co) {
-  return 'customOption_' + co.option_id
-}
-
-function _getSelectedOptionValue (optionType, optionValues = [], inputValue) {
-  switch (optionType) {
-    case 'field': {
-      return inputValue
-    }
-    case 'radio':
-    case 'select':
-    case 'drop_down': {
-      const selectedValue = optionValues.find((value) => value.option_type_id === inputValue) || {}
-
-      return selectedValue.option_type_id || ''
-    }
-    case 'checkbox': {
-      return optionValues.filter((value) => (inputValue || []).includes(value.option_type_id))
-        .map((value) => value.option_type_id)
-        .join(',')
-    }
-    default: {
-      return ''
-    }
-  }
-}
 
 export const ProductCustomOptions = {
   name: 'ProductCustomOptions',
@@ -70,8 +30,8 @@ export const ProductCustomOptions = {
       }
 
       return customOptions.reduce((selectedOptions, option) => {
-        const fieldName = _fieldName(option)
-        selectedOptions[fieldName] = _getSelectedOptionValue(option.type, option.values, this.inputValues[fieldName])
+        const fieldName = customOptionFieldName(option)
+        selectedOptions[fieldName] = selectedCustomOptionValue(option.type, option.values, this.inputValues[fieldName])
         return selectedOptions
       }, {})
     }
@@ -93,8 +53,8 @@ export const ProductCustomOptions = {
     }),
     setupInputFields () {
       for (const customOption of this.product.custom_options) {
-        const fieldName = _fieldName(customOption)
-        this['inputValues'][fieldName] = _defaultOptionValue(customOption)
+        const fieldName = customOptionFieldName(customOption)
+        this['inputValues'][fieldName] = defaultCustomOptionValue(customOption)
         if (customOption.is_require) { // validation rules are very basic
           this.validation.rules[fieldName] = 'required' // TODO: add custom validators for the custom options
         }
@@ -102,14 +62,14 @@ export const ProductCustomOptions = {
       }
     },
     optionChanged (option) {
-      const fieldName = _fieldName(option)
+      const fieldName = customOptionFieldName(option)
       this.validateField(option)
       this.setCustomOptionValue({ optionId: option.option_id, optionValue: this.selectedOptions[fieldName] })
       this.$store.dispatch('product/setCustomOptions', { product: this.product, customOptions: this.$store.state.product.current_custom_options }) // TODO: move it to "AddToCart"
       this.$bus.$emit('product-after-customoptions', { product: this.product, option: option, optionValues: this.selectedOptions })
     },
     validateField (option) {
-      const fieldName = _fieldName(option)
+      const fieldName = customOptionFieldName(option)
       const validationRule = this.validation.rules[fieldName]
       this.product.errors.custom_options = null
       if (validationRule) {
